core process and initial crud endpoints

+++ src/main/java/br/ufsc/tcc/gamifyEngine/dao/AttributeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.ufsc.tcc.gamifyEngine.model.Attribute;
 
-public interface AttributeDao  extends CrudRepository<Attribute, Long> {
+public interface AttributeDao  extends CrudRepository<Attribute, Integer> {
 	public Attribute findById(int attributeId);
 }
