+++ src/main/java/br/ufsc/tcc/gamifyEngine/dao/BadgeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.ufsc.tcc.gamifyEngine.model.Badge;
 
-public interface BadgeDao  extends CrudRepository<Badge, Long> {
+public interface BadgeDao  extends CrudRepository<Badge, Integer> {
 	public Badge findById(int badgeId);
 }
